requirements.txt
asgiref==3.6.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
Django==4.1.7
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
fonttools==4.39.2
idna==3.4
imageio==2.26.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.2
MarkupSafe==2.1.2
matplotlib==3.7.1
murmurhash==1.0.9
networkx==3.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
PassportEye==2.2.1
pathy==0.10.1
pdfminer==20191125
Pillow==9.4.0
preshed==3.0.8
psycopg2==2.9.5
pycryptodome==3.17
pydantic==1.10.7
PyMuPDF==1.21.1
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
thinc==8.1.9
threadpoolctl==3.1.0
tifffile==2023.3.21
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wasabi==1.1.1
